AMP-30948 : Add  Importer progress page to context

+++ amp/WEB-INF/src/org/digijava/module/aim/action/dataimporter/model/ImportDataModel.java
@@ -49,7 +49,7 @@ public class ImportDataModel {
     private Object iati_last_update_date;
     private Object approved_by;
     private Object approval_date;
-    private Long approval_status;
+    private Integer approval_status;
     private Object archived;
     private Set<Object> indicators;
     private Set<Object> activity_documents;
@@ -377,11 +377,11 @@ public class ImportDataModel {
         this.approval_date = approval_date;
     }
 
-    public Long getApproval_status() {
+    public Integer getApproval_status() {
         return approval_status;
     }
 
-    public void setApproval_status(Long approval_status) {
+    public void setApproval_status(Integer approval_status) {
         this.approval_status = approval_status;
     }
